Refactor backend text + image code
