+++ digiprod_gen/backend_api/image/caption.py
@@ -3,7 +3,7 @@ import requests
 
 from typing import List
 from PIL import Image
-from digiprod_gen.backend.image.conversion import pil2bytes_io, bytes2pil
+from digiprod_gen.backend_api.image.conversion import pil2bytes_io, bytes2pil
 from digiprod_gen.backend_api.models.mba import MBAProduct
 from digiprod_gen.backend_api.models.common import ImageCaptioningModel
 from digiprod_gen.backend_api.utils import booleanize
