+++ digiprod_gen/backend_api/image/compress.py
