+++ digiprod_gen/backend_api/image/stabilityai.py
