+++ digiprod_gen/backend_api/image/utils.py
