+++ digiprod_gen/frontend/sidebar.py
@@ -10,7 +10,7 @@ from digiprod_gen.frontend.tab.crawling.tab_crawling import crawl_mba_overview_a
 from digiprod_gen.frontend.tab.image_generation.selected_products import crawl_details_update_overview_page
 from digiprod_gen.frontend.tab.image_generation.prompt_generation import prompt_generation
 from digiprod_gen.frontend.tab.upload.listing_generation import listing_generation
-from digiprod_gen.backend.image.caption import extend_mba_products_with_caption
+from digiprod_gen.backend_api.image.caption import extend_mba_products_with_caption
 from digiprod_gen.frontend.session import read_session
 from digiprod_gen.backend_api.utils.utils import booleanize
