+++ digiprod_gen/frontend/tab/image_generation/image_generation.py
@@ -3,7 +3,7 @@ import streamlit as st
 from PIL import Image
 from digiprod_gen.backend_api.models.session import ImageGenData
 from digiprod_gen.backend_api.models.common import ImageGenerationModel
-from digiprod_gen.backend.image import generation
+from digiprod_gen.backend_api.image import generation
 
 def update_session_selected_prompt(session_image_gen_data: ImageGenData, image_gen_prompt=None):
     session_image_gen_data.image_gen_prompt_selected = image_gen_prompt if image_gen_prompt else st.session_state["image_gen_prompt"]
