+++ digiprod_gen/frontend/tab/upload/listing_generation.py
@@ -4,13 +4,13 @@ import streamlit as st
 from streamlit.delta_generator import DeltaGenerator
 from langchain.chat_models import ChatOpenAI
 
-from digiprod_gen.backend.generative_ai.text.text_gen_fns import get_product_text_gen, remove_banned_words
+from digiprod_gen.backend_api.text.text_gen_fns import get_product_text_gen, remove_banned_words
 from digiprod_gen.backend.prompt_engineering.utils import extract_list_from_output
 from digiprod_gen.frontend.session import read_session
 
 from digiprod_gen.backend_api.models.mba import CrawlingMBARequest, MBAProductTextType, MBAMarketplaceDomain
 from digiprod_gen.backend_api.models.session import SessionState
-from digiprod_gen.backend.generative_ai.text.mba_banned_word import MBA_BANNED_WORDS
+from digiprod_gen.backend_api.text.mba_banned_word import MBA_BANNED_WORDS
 
 
 def listing_generation(st_tab_ig: DeltaGenerator):
