+++ digiprod_gen/frontend/tab/upload/views.py
@@ -1,8 +1,8 @@
 import streamlit as st
 from typing import List, Tuple
 
-from digiprod_gen.backend.image import conversion as img_conversion
-from digiprod_gen.backend.image.utils import hex_to_rgba
+from digiprod_gen.backend_api.image import conversion as img_conversion
+from digiprod_gen.backend_api.image.utils import hex_to_rgba
 from digiprod_gen.backend_api.utils.helper import Timer
 from digiprod_gen.backend_api.models.session import MBAUploadData, SessionState, ImageGenData, DigiProdGenStatus
 from digiprod_gen.backend_api.models.mba import MBAMarketplaceDomain, MBAProductCategory, MBAProductColor, \
